fix(game): settle raid outcomes on effects, not just the enum
C1: an extracted player could still be shot, and the kill wiped the haul.
Exclusivity was airtight on ESarkoRaidOutcome but not on the effects: the
second FinishRaid(Died) was refused while HandleDeath's ClearOnDeath() still
ran, so the summary read EXTRACTED over an empty backpack and Task 8 would
submit an empty result. Three layers now:
  1. USarkoHealthComponent::ApplyDamage takes an earliest return once the
     raid has an outcome (one game-state lookup, server-only path, null-world
     tolerant like the authority guard, with a SetRaidFinishedForTest seam).
  2. ASarkoAIController::Tick returns early on a finished raid, so the bots
     visibly stand down instead of tracking a frozen pawn.
  3. ASarkoCharacter::HandleDeath refuses every death side-effect when the
     outcome is already settled, and logs — a late death must never be able
     to corrupt a decided result.
Pinned by Sarko.Extract.SettledOutcomeTakesNoMoreDamage.

I1: FinishRaid now clears the backpack for every non-Extracted outcome,
before it writes Outcome. MIA had no death path, so spec §4.5's "MIA is
death, loot lost" was false and the MIA summary itemised loot the player had
just lost. The rule is the pure SarkoRaid::OutcomeLosesHaul, pinned by
Sarko.Extract.LosingOutcomesLoseTheHaul, which makes the HUD's "the server
emptied the backpack before the outcome was set" a fact.

m1: remaining per-frame HUD allocations — the clock Printf is cached on the
integer second, the ammo readout on the magazine count, the backpack readout
on used/max, and the interact label's GetTextSize is hoisted like
CachedReloadingWidth.

m2: FreezeForRaidEnd clears ExtractZoneIndex/ExtractDwellSeconds, so the
dwell chip stops drawing under the summary.

m3: LootSalt widened to int64 (two of the raid GUID's words, both halves
folded into ContainerSeed at different stages). 32 bits was offline-sweepable
from one observed roll, which reopened the whole raid; the comment now states
the honest bound rather than implying strength.

m4: RegisteredContainers' header comment no longer claims registration order,
which RemoveAtSwap does not preserve.

Verified: Scripts/run-tests.sh — 49 tests performed, 0 failed.

+++ SarkoGame/Source/SarkoGame/Core/SarkoRaidGameMode.h
@@ -82,9 +82,11 @@ public:
 	 *
 	 * Not derived from Seed, and not seeded from the raid clock: both are things a
 	 * client knows. Regenerated per raid, so nothing learned from one raid carries
-	 * into the next.
+	 * into the next — which is also the bound on how strong this is, since 64 bits
+	 * of salt raise the cost of an offline sweep from one observed roll to a few
+	 * rather than making it impossible (see SarkoLoot::ContainerSeed).
 	 */
-	int32 LootSalt = 0;
+	int64 LootSalt = 0;
 
 	/** The layout this raid was loaded from; pawns spawn against it. */
 	FSarkoMapLayout CachedLayout;
